Migrate to Vite SPA hosted on Cloudflare Workers static assets
Replaces the Next.js server app (hosted on Render) with a fully static
Vite + React SPA served from Cloudflare Workers static assets:

- Move /api/get-domains logic client-side (src/lib/get-domains.ts);
  dRPC + subgraph keys are now domain-restricted frontend keys
- Replace Next.js with Vite: index.html + src/main.tsx entry,
  page.tsx -> App.tsx, next/image -> img, next/font -> Google Fonts link
- Rename env vars NEXT_PUBLIC_* -> VITE_* (import.meta.env)
- Add wrangler.jsonc serving ./dist with SPA fallback
- Build: tsc --noEmit && vite build (~1.5s); deploy via wrangler deploy
