+++ example.env
@@ -1,6 +1,5 @@
-NEXT_PUBLIC_WALLET_CONNECT_PROJECT_ID=
-NEXT_PUBLIC_DRPC_API_KEY_FRONTEND=
-DRPC_API_KEY_BACKEND=
-SUBGRAPH_URL=
-SEPOLIA_SUBGRAPH_URL=
-NEXT_PUBLIC_FACTORY_ADDRESS=0xDddddDdDDD8Aa1f237b4fa0669cb46892346d22d
+VITE_WALLET_CONNECT_PROJECT_ID=
+VITE_DRPC_API_KEY_FRONTEND=
+VITE_SUBGRAPH_URL=
+VITE_SEPOLIA_SUBGRAPH_URL=
+VITE_FACTORY_ADDRESS=0xDddddDdDDD8Aa1f237b4fa0669cb46892346d22d
