+++ index.html
@@ -1,0 +1,55 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="icon" href="/favicon.ico" />
+
+    <title>Durin - ENS L2 Names Made Easy</title>
+    <meta
+      name="description"
+      content="The developer's gateway to ENS L2 Subnames"
+    />
+
+    <!-- Open Graph -->
+    <meta property="og:type" content="website" />
+    <meta property="og:url" content="https://durin.dev" />
+    <meta property="og:title" content="Durin - ENS L2 Names Made Easy" />
+    <meta
+      property="og:description"
+      content="The developer's gateway to ENS L2 Subnames"
+    />
+    <meta property="og:image" content="https://durin.dev/opengraph-image.jpg" />
+    <meta property="og:image:width" content="1200" />
+    <meta property="og:image:height" content="630" />
+    <meta property="og:image:alt" content="Durin - L2 Names Made Easy" />
+
+    <!-- Twitter -->
+    <meta name="twitter:card" content="summary_large_image" />
+    <meta name="twitter:title" content="Durin - L2 Names Made Easy" />
+    <meta
+      name="twitter:description"
+      content="The developer's gateway to ENS L2 Subnames"
+    />
+    <meta
+      name="twitter:image"
+      content="https://durin.dev/opengraph-image.jpg"
+    />
+
+    <!-- Gelasio (previously loaded via next/font/google) -->
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link
+      rel="preconnect"
+      href="https://fonts.gstatic.com"
+      crossorigin
+    />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Gelasio:wght@400;500;700&display=swap"
+      rel="stylesheet"
+    />
+  </head>
+  <body class="font-gelasio">
+    <div id="root"></div>
+    <script type="module" src="/src/main.tsx"></script>
+  </body>
+</html>
