--- next.config.ts
@@ -1,15 +1,0 @@
-import type { NextConfig } from "next";
-
-const nextConfig: NextConfig = {
-  webpack: (config) => {
-    config.resolve.fallback = {
-      ...config.resolve.fallback,
-      encoding: false,
-      bufferutil: false,
-      "utf-8-validate": false,
-    };
-    return config;
-  },
-};
-
-export default nextConfig;
