+++ package.json
@@ -3,19 +3,19 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
+    "dev": "vite",
+    "build": "tsc --noEmit && vite build",
+    "preview": "npm run build && wrangler dev",
+    "deploy": "npm run build && wrangler deploy"
   },
   "dependencies": {
     "@ensdomains/ensjs": "^4.1.2",
     "@rainbow-me/rainbowkit": "^2.2.4",
     "@tanstack/react-query": "^5.67.2",
+    "clsx": "^2.1.1",
     "evm-providers": "^0.6.0",
     "lucide": "^0.456.0",
     "lucide-react": "^0.456.0",
-    "next": "^15.0.2",
     "pino-pretty": "^12.1.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
@@ -28,10 +28,11 @@
     "@types/node": "^24.5.2",
     "@types/react": "^18",
     "@types/react-dom": "^18",
-    "eslint": "^8",
-    "eslint-config-next": "15.0.2",
+    "@vitejs/plugin-react": "^6.0.3",
     "postcss": "^8",
     "tailwindcss": "^3.4.1",
-    "typescript": "^5.9.2"
+    "typescript": "^5.9.2",
+    "vite": "^8.1.5",
+    "wrangler": "^4.107.1"
   }
 }
