+++ src/App.tsx
@@ -1,7 +1,3 @@
-"use client";
-import Image from "next/image";
-import { Gelasio } from "next/font/google";
-import Link from "next/link";
 import { ConnectButton } from "@rainbow-me/rainbowkit";
 import { useState, useEffect } from "react";
 import { useAccount } from "wagmi";
@@ -9,15 +5,11 @@ import DeployButton from "./components/deploy-button";
 import { type Address } from "viem";
 import UpdateResolverButton from "./components/update-resolver-button";
 import SetRegistryButton from "./components/set-registry-button";
-import { Domain } from "../lib/types";
+import { Domain } from "./lib/types";
+import { getDomains } from "./lib/get-domains";
 import { Copy, Check, ExternalLink, ScrollText, Sparkles } from "lucide-react";
 
 import { RESOLVER_ADDRESSES } from "@/lib/utils";
-
-const gelasio = Gelasio({
-  weight: ["500", "400", "700"],
-  subsets: ["latin"],
-});
 
 interface Transactionn {
   action: string;
@@ -90,12 +82,12 @@ export default function Home() {
   return (
     <div className="flex flex-col h-screen font-sans text-stone-900 relative">
       <div className="flex items-center justify-between h-16 px-4 md:px-10 mt-4">
-        <Image
+        <img
           alt="logo"
-          src="durin-logo-with-arch.svg"
+          src="/durin-logo-with-arch.svg"
           width={115}
           height={30}
-        ></Image>
+        />
 
         <ConnectButton
           showBalance={false}
@@ -109,23 +101,23 @@ export default function Home() {
 
       {/* Main Content */}
       <main className="flex px-4 lg:px-0 flex-col flex-grow w-full max-w-5xl gap-6 mx-auto">
-        <Image
+        <img
           className="rounded-lg mt-4 hidden md:block"
           src="/banner.webp"
           alt="durin"
           width={1024}
           height={1024}
-        ></Image>
-        <Image
+        />
+        <img
           className="rounded-lg mt-4 self-center block md:hidden"
           src="/banner-square.webp"
           alt="durin"
           width={256}
           height={256}
-        ></Image>
+        />
         <div className="relative">
           <h1
-            className={`self-start ${gelasio.className} w-full md:w-[calc(100%-340px)] text-2xl md:text-3xl`}
+            className={`self-start font-gelasio w-full md:w-[calc(100%-340px)] text-2xl md:text-3xl`}
           >
             Issue onchain ENS subdomains on an L2
           </h1>
@@ -133,13 +125,13 @@ export default function Home() {
             <div className="self-start mt-4 w-full md:w-[calc(100%-340px)] text-stone-600 mr-12">
               Durin is an opinionated approach to issuing ENS L2 subdomains.
               Durin is{" "}
-              <Link
+              <a
                 target="_blank"
                 href={"https://github.com/namestonehq/durin"}
                 className="underline underline-offset-4"
               >
                 open source
-              </Link>{" "}
+              </a>{" "}
               and PRs are welcomed.
             </div>
             <div className="absolute hidden md:block right-4 -top-20 bg-stone-150 p-6 w-80 rounded-lg shadow-md">
@@ -161,7 +153,7 @@ export default function Home() {
           </div>
           <hr className="bg-stone-100 my-6" />
           {/* Deploy L2 Registry */}
-          <h2 className={`${gelasio.className} mb-3 text-xl`}>
+          <h2 className={`font-gelasio mb-3 text-xl`}>
             1. Deploy the L2 Registry
           </h2>
           <div className="flex flex-col md:flex-row gap-20">
@@ -255,7 +247,7 @@ export default function Home() {
                       className="flex items-center justify-between w-full p-3 text-sm bg-gray-100 rounded"
                     >
                       <div className="flex items-center gap-2">
-                        <Image
+                        <img
                           src={`/${chainName
                             .toLowerCase()
                             .replace(/\s+/g, "")}-icon.svg`}
@@ -266,7 +258,7 @@ export default function Home() {
                         />
                         <span>{getFullChainName(chainName, isTestnet)}</span>
                       </div>
-                      <Image
+                      <img
                         src="/chevron-down.svg"
                         alt="chevron"
                         width={16}
@@ -299,7 +291,7 @@ export default function Home() {
                               chainName === chain.name ? "bg-gray-50" : ""
                             }`}
                           >
-                            <Image
+                            <img
                               src={`/${chain.icon}-icon.svg`}
                               alt={chain.name}
                               width={24}
@@ -316,7 +308,7 @@ export default function Home() {
                   </div>
                   <div className="font-light mt-6">
                     Deploy L2{" "}
-                    <Link
+                    <a
                       target="_blank"
                       className="underline underline-offset-4"
                       href={
@@ -324,7 +316,7 @@ export default function Home() {
                       }
                     >
                       Registry
-                    </Link>{" "}
+                    </a>{" "}
                     on {getFullChainName(chainName, isTestnet)}
                   </div>
                   <div className="text-sm text-stone-600 ">
@@ -366,8 +358,8 @@ export default function Home() {
                     </button>
                   )}
                   {registryAddress && (
-                    <Image
-                      src={"etherscan-logo.svg"}
+                    <img
+                      src={"/etherscan-logo.svg"}
                       alt="etherscan"
                       width={24}
                       height={24}
@@ -392,7 +384,7 @@ export default function Home() {
             </div>
           </div>
           {/* Configure L1 Resolver */}
-          <h2 className={`${gelasio.className} mb-3 mt-12 text-xl`}>
+          <h2 className={`font-gelasio mb-3 mt-12 text-xl`}>
             2. Configure L1 Resolver
           </h2>
           <div className="flex flex-col md:flex-row gap-20">
@@ -452,8 +444,8 @@ export default function Home() {
                       network as keyof typeof RESOLVER_ADDRESSES
                     ]?.slice(-4)}
                   </div>
-                  <Image
-                    src={"etherscan-logo.svg"}
+                  <img
+                    src={"/etherscan-logo.svg"}
                     alt="etherscan"
                     width={24}
                     height={24}
@@ -516,7 +508,7 @@ export default function Home() {
             </div>
           </div>
           {/* Customize Registrar */}
-          <h2 className={`${gelasio.className} mb-3 mt-12 text-xl`}>
+          <h2 className={`font-gelasio mb-3 mt-12 text-xl`}>
             3. Customize Registrar
           </h2>
           <div className="flex flex-col md:flex-row gap-20">
@@ -545,13 +537,13 @@ export default function Home() {
                     </div>
                     <div className="text-sm mt-2 text-stone-500">
                       Select a template{" "}
-                      <Link
+                      <a
                         href="https://github.com/namestonehq/durin/blob/main/src/examples/L2Registrar.sol"
                         target="_blank"
                         className="underline underline-offset-4"
                       >
                         L2 registrar
-                      </Link>{" "}
+                      </a>{" "}
                       and modify it as needed. You will need your registry
                       address.
                     </div>
@@ -566,12 +558,12 @@ export default function Home() {
                         }}
                         className="flex items-center gap-2 h-9 px-2 text-sm border rounded-lg shadow text- text-stone-900 hover:bg-stone-100"
                       >
-                        <Image
+                        <img
                           alt="github"
                           src="/github.svg"
                           width={16}
                           height={16}
-                        ></Image>
+                        />
                         Customize
                       </button>
                     </div>
@@ -604,8 +596,8 @@ export default function Home() {
                     </button>
                   )}
                   {registryAddress && (
-                    <Image
-                      src={"etherscan-logo.svg"}
+                    <img
+                      src={"/etherscan-logo.svg"}
                       alt="etherscan"
                       width={24}
                       height={24}
@@ -630,7 +622,7 @@ export default function Home() {
             </div>
           </div>
           {/* Connect Registrar to Registry*/}
-          <h2 className={`${gelasio.className} mb-3 mt-12 text-xl`}>
+          <h2 className={`font-gelasio mb-3 mt-12 text-xl`}>
             4. Connect Registrar to Registry
           </h2>
           <div className="flex flex-col md:flex-row gap-20">
@@ -666,7 +658,7 @@ export default function Home() {
               </div>
             </div>
           </div>
-          <h2 className={`${gelasio.className} mb-3 mt-12 text-xl`}>
+          <h2 className={`font-gelasio mb-3 mt-12 text-xl`}>
             5. Mint your first subdomain
           </h2>
           <div className="flex flex-col md:flex-row gap-20">
@@ -694,7 +686,7 @@ export default function Home() {
                       Etherscan. <br />
                       <br />
                       Under Contract select register() and write.
-                      <Image
+                      <img
                         src="/etherscan-screenshot.webp"
                         alt="etherscan"
                         width={400}
@@ -711,7 +703,7 @@ export default function Home() {
         {/* Tx history */}
         <div className="items-start w-full flex-col gap-2">
           <div className="flex items-center justify-between">
-            <div className={`${gelasio.className} text-xl py-4`}>
+            <div className={`font-gelasio text-xl py-4`}>
               Transaction History
             </div>
             <div className="relative group">
@@ -773,13 +765,13 @@ export default function Home() {
           <div className="flex  items-center justify-center sm:w-[800px] mx-auto">
             <span className="text-sm text-neutral-500">
               Originally created by{" "}
-              <Link
+              <a
                 href="https://x.com/AlexSlobodnik"
                 target="_blank"
                 className="text-neutral-300 hover:text-orange-400 transition-colors duration-300 ease-in-out"
               >
                 Alex Slobodnik
-              </Link>
+              </a>
             </span>
           </div>
         </div>
@@ -815,21 +807,8 @@ function DomainSelector({
 
       try {
         setIsLoading(true);
-        const baseUrl =
-          process.env.NEXT_PUBLIC_BASE_URL || window.location.origin;
-        const url = new URL(`${baseUrl}/api/get-domains`);
-        url.searchParams.append("address", address);
-        url.searchParams.append("network", network);
-
-        const response = await fetch(url.toString());
-        const data: Domain[] = await response.json();
-
-        if (response.status === 200) {
-          setUserDomains(data.filter((name) => name !== null));
-        } else {
-          console.error("Error fetching domains:", data);
-          setUserDomains([]);
-        }
+        const data = (await getDomains(address, network)) as Domain[];
+        setUserDomains(data.filter((name) => name !== null));
       } catch (error) {
         console.error("Error fetching ENS names:", error);
         setUserDomains([]);
@@ -883,7 +862,7 @@ function DomainSelector({
               : ""
           }`}
         />
-        <Image
+        <img
           alt="chevron"
           src="/chevron-down.svg"
           width={16}
