+++ src/components/custom-connect-button.tsx
@@ -1,8 +1,5 @@
-"use client";
-
 import { ConnectButton } from "@rainbow-me/rainbowkit";
 import { cn, formatEthereumAddress } from "@/lib/utils";
-import Image from "next/image";
 
 export function CustomConnectButton({ className }: { className?: string }) {
   return (
@@ -93,7 +90,7 @@ export function CustomConnectButton({ className }: { className?: string }) {
                     "text-white transition-colors duration-300 rounded-lg bg-stone-900 w-fit hover:bg-stone-700 disabled:opacity-50 flex items-center justify-center gap-2"
                   )}
                 >
-                  <Image
+                  <img
                     src="/logo-small-dark.svg"
                     alt="gon.id Logo"
                     width={20}
