+++ src/components/set-registry-button.tsx
