+++ src/index.css
@@ -19,3 +19,7 @@ body {
   background: var(--background);
   font-family: Arial, Helvetica, sans-serif;
 }
+
+.font-gelasio {
+  font-family: "Gelasio", Georgia, "Times New Roman", serif;
+}
